+++ src/pages/reservation/components/ThirdStep.jsx
@@ -1,0 +1,14 @@
+import { Link } from 'react-router-dom';
+
+const ThirdStep = ({ list, setStep }) => {
+  return (
+    <div>
+      third step
+      <Link to='/'>
+        <button className='btn'>확인</button>
+      </Link>
+    </div>
+  );
+};
+
+export default ThirdStep;
